[ts] [stock_counter] [03_add_factory_method] Add Factory Method
