+++ examples/ts/ts-stock_counter-03_add_factory_method/src/FeedParser.ts
@@ -1,0 +1,6 @@
+import {Feed} from "./Feed";
+import {Flower} from "./Flower";
+
+export class FeedParser {
+    parse: (feed :Feed) => Array<Flower>
+ }
